+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Custom toolbar with Fullscreen tools [#56](https://github.com/interalia-studio/fullscreen/pull/56)
+- Use stitches for CSS-in-JS [#56](https://github.com/interalia-studio/fullscreen/pull/56)
+
+### Fixed
+
+- Fix release action running on every push to `main` branch [#56](https://github.com/interalia-studio/fullscreen/pull/56)
+
 ## [0.1.0]
 
 ### Added